Modified Tests as suggested.
