+++ test/ProfilingPlugin.unittest.js
@@ -1,9 +1,6 @@
 "use strict";
 
 const ProfilingPlugin = require("../lib/debug/ProfilingPlugin");
-const path = require("path");
-const fs = require("fs");
-const webpack = require("../");
 
 describe("Profiling Plugin", () => {
 	it("should persist the passed outpath", () => {
@@ -16,26 +13,6 @@ describe("Profiling Plugin", () => {
 	it("should handle no options", () => {
 		const plugin = new ProfilingPlugin();
 		expect(plugin.outputPath).toBe("events.json");
-	});
-
-	it("should handle outpath with folder", done => {
-		const finalPath = "test/fixtures/profilingPath/events.json";
-		const compiler = webpack({
-			context: "/",
-			entry: "./fixtures/a.js",
-			plugins: [
-				new webpack.debug.ProfilingPlugin({
-					outputPath: finalPath
-				})
-			]
-		});
-		const outputPath = path.join(__dirname, "/fixtures/profilingPath");
-		compiler.run(err => {
-			if (err) return done(err);
-			if (!fs.existsSync(outputPath))
-				return done(new Error("Folder should be created."));
-			done();
-		});
 	});
 
 	it("should handle when unable to require the inspector", () => {
